requirements_streamlit.txt
streamlit>=1.25.0
streamlit-chat>=0.1.1
Pillow>=9.0.0
langchain>=0.0.312
openai>=1.3.0
chromadb>=0.4.10
sentence-transformers>=2.2.0
transformers>=4.35.0
torch>=2.0.0
pandas>=2.0.0
tiktoken>=0.5.0
markdown>=3.4.0
beautifulsoup4>=4.11.0
fastapi>=0.100.0
python-dotenv>=1.0.0